+++ src/main/java/com/jingxu/shopdemo/utils/UserContext.java
@@ -2,14 +2,14 @@ package com.jingxu.shopdemo.utils;
 
 public class UserContext {
     // 保存用户信息的 ThreadLocal
-    private static final ThreadLocal<Integer> currentUser = new ThreadLocal<>();
+    private static final ThreadLocal<Long> currentUser = new ThreadLocal<>();
     // 设置用户信息
-    public static void set(Integer userId) {
+    public static void set(Long userId) {
         currentUser.set(userId);
     }
 
     // 获取用户信息
-    public static Integer get() {
+    public static Long get() {
         return currentUser.get();
     }
